# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STEEP/PESTLE Foresight Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trends:
    get:
      operationId: listTrends
      tags: [trends]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Trend" }
    post:
      operationId: createTrend
      tags: [trends]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTrendBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trend" }
  /trends/{id}:
    get:
      operationId: getTrend
      tags: [trends]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trend" }
    patch:
      operationId: updateTrend
      tags: [trends]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTrendBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trend" }

  # ---------- Collections ----------
...
components:
  schemas:
...
    Trend:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        claim: { type: string }
        rationale: { type: string }
        category: { type: string }
        timeHorizon: { type: string }
        confidenceRating: { type: integer }
        impactRating: { type: integer }
        status:
          type: string
          enum: [preliminary, established, contested]
        supportingSignalIds:
          type: array
          items: { type: string }
        contradictingSignalIds:
          type: array
          items: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, title, claim, rationale, category, timeHorizon, confidenceRating, impactRating, status, supportingSignalIds, contradictingSignalIds, authorId, authorName, createdAt, updatedAt]
    CreateTrendBody:
      type: object
      properties:
        title: { type: string }
        claim: { type: string }
        rationale: { type: string }
        category: { type: string }
        timeHorizon: { type: string }
        confidenceRating: { type: integer }
        impactRating: { type: integer }
        supportingSignalIds:
          type: array
          items: { type: string }
        contradictingSignalIds:
          type: array
          items: { type: string }
      required: [title, claim, category, supportingSignalIds]
    UpdateTrendBody:
      type: object
      properties:
        title: { type: string }
        claim: { type: string }
        rationale: { type: string }
        category: { type: string }
        timeHorizon: { type: string }
        confidenceRating: { type: integer }
        impactRating: { type: integer }
        status:
          type: string
          enum: [preliminary, established, contested]
        supportingSignalIds:
          type: array
          items: { type: string }
        contradictingSignalIds:
          type: array
          items: { type: string }

    # ----- Collection -----
...